requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
html5lib==0.999999999
idna==2.6
peewee==2.10.1
PyMySQL==0.7.11
python-dateutil==2.6.1
requests==2.18.4
six==1.10.0
urllib3==1.22
webencodings==0.5.1
